requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.39.0
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.2
beautifulsoup4==4.12.3
python-docx==1.1.2
python-jobspy==1.1.82
